Reworked path list management.
Now by the time the path_list code gets an entry to add, there's
no interpretation to be done, only duplicate pruning.
Everything coming in is funnelled through add_entry (called from
main() for each command line arg), which checks for whether it needs
shortname expansion or path splitting and calls appropriate functions.

Path splitting is now in the path_list code, and returns a new
path_list with the components of the path it's given to split.

remove_entry and remove_entries are also implemented.  They are
expected to be used in the next commit.