Added capability to remove entries from path being constructed.
To remove an entry, use '^entry' on the command line.  "entry" is
expanded by shortname expansion and path splitting if appropriate,
then the set of path_entries it expands to is removed from the
current path.

Command line args are handled in order, so 'showpath ...stuff... ^foo foo'
will force foo to the end of the path (build path, remove foo if
it's there, then add it at the end).