Fixed args to expand_home
Now takes (from,to,maxlen) instead of (src,dest).

from/to order no longer matches strcpy et al, but since this is
munging and not an exact copy, the reading 'expand from to to' makes
sense.  It also means that related args ("buffer pointers" and
independently "related to output") are together.

Adding maxlen instead of depending on the sizes of the allocated
buffers from callers is a more necessary change.  Now we don't have
to worry about maintenance skew that doesn't update everything in
sync.
(Dynamically allocating appropriate amounts of memory would also
be Really Nice.)